% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lidar}
\alias{lidar}
\title{LIDAR data}
\arguments{
\item{range}{distance travelled before the light is reflected back to its source.}

\item{logratio}{logarithm of the ratio of received light from two laser sources.}
}
\description{
The lidar data frame has 221 observations from a light detection and ranging (LIDAR) experiment.

This data set was part of the package SemiPar, which is no longer available at CRAN in compiled form.
}
\keyword{data}
