% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p8105.datasets-package.r
\docType{package}
\name{tidyfun}
\alias{tidyfun}
\alias{tidyfun-package}
\title{p8105.datasets}
\description{
`p8105.datasets` is an R package containing datasets for P8105.
}
\details{
The package includes

- AirBnB data
  - Instacart data
  - NOAA data
  - NYC Restaurant Inspection data
}
