% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{weather_df}
\alias{weather_df}
\title{Weather}
\arguments{
\item{name}{name of the weather station}

\item{id}{weather station ID number}

\item{date}{date of observation}

\item{prcp}{total daily precipitation}

\item{tmax}{maximum daily temperature}

\item{tmin}{minimum daily temperature}
}
\description{
The lidar data frame has 2,190 total observations, coming from two years of monitoring at three US weather stations.

These data were originally accessed using the `rnoaa` package, which accessed NOAA data using an API. That package has been discontinued.
}
\keyword{data}
